Broken things

+++ libft/ft_putnbr.c
@@ -6,13 +6,13 @@
 /*   By: sescolas <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/01/09 20:31:11 by sescolas          #+#    #+#             */
-/*   Updated: 2017/01/20 10:34:04 by sescolas         ###   ########.fr       */
+/*   Updated: 2017/08/08 15:41:29 by sescolas         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	ft_putnbr(int n)
+void	ft_putnbr(ssize_t n)
 {
 	if (n < 0)
 	{
